feat: #102. Рефакторинг TelegramCommandHandler
