+++ NotificationBot/Interfaces/TelegramCommands/ICallbackQueryService.cs
@@ -1,0 +1,9 @@
+﻿using Telegram.Bot.Types;
+
+namespace NotificationsBot.Interfaces.TelegramCallback
+{
+    public interface ICallbackQueryService
+    {
+        Task HandleCallbackQuery(CallbackQuery callbackQuery, Message message);
+    }
+}
